chore: fix typing with new highspy version (#795)
* fix(highs): pass uint8 integrality array to changeColsIntegrality

highspy's changeColsIntegrality expects var_types as a uint8 array
(matching HiGHS's uint8-backed HighsVarType enum). The persistent
update path already builds it as uint8 (_apply_var_types); align the
direct-solve path so it matches highspy's tightened type stubs and is
robust across versions.


* fix(highs): silence assignment error on highspy lazy module

highspy is imported under TYPE_CHECKING, so mypy types `highspy` as the
module; assigning the `_LazyModule` shim then trips [assignment] once
highspy ships recognized types (highspy 1.13.1). Mirror the existing
gurobipy line with `# type: ignore[assignment]`.


---------

+++ linopy/solvers.py
@@ -294,7 +294,7 @@ class _LazyModule:
 
 
 gurobipy = _LazyModule("gurobipy")  # type: ignore[assignment]
-highspy = _LazyModule("highspy")
+highspy = _LazyModule("highspy")  # type: ignore[assignment]
 scip = _LazyModule("pyscipopt")
 cplex = _LazyModule("cplex")
 knitro = _LazyModule("knitro")
@@ -1656,7 +1656,7 @@ class Highs(Solver[None]):
             int_mask = (vtypes == "B") | (vtypes == "I") | (vtypes == "S")
             labels = np.arange(len(vtypes))[int_mask]
             integrality = np.array(
-                [integrality_map[v] for v in vtypes[int_mask]], dtype=np.int32
+                [integrality_map[v] for v in vtypes[int_mask]], dtype=np.uint8
             )
             h.changeColsIntegrality(len(labels), labels, integrality)
